+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/CHANGELOG.md
@@ -144,6 +144,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - In the ACLs and CCLs, display fields for both the `Root` type, and the `QueryRoot`+`MutationRoot` types ([#989](https://github.com/leoloso/PoP/pull/989))
 - The entities from the WordPress data model are not namespaced anymore ([#990](https://github.com/leoloso/PoP/pull/990))
 - Prettyprint GraphQL queries in the module docs
+- Exposed the `__schema` field in the ACLs, allowing us to disable introspection
 
 ### Fixed